Add option to override "not implemented" status (#512)
We have a bunch of cards that are implemented in the BE but are missing
critical pieces in the FE in order to be able to function. Added a way
to override the "not implemented" status manually so that testers can
avoid these cards.

Updated as many cards as I'm aware of to have this marker.