Add optional prompt support for "basic" lookat / reveal case (#506)
Beginning the process of refactoring for every lookat / reveal effect to
use the display prompt instead of relying on text messages. Since there
are a lot of cards that need to be updated and many of them are special
cases, we are adding an optional property that enables the prompt in the
`lookAt()` and `reveal()` functions. This will be removed once all cards
are ported over.

Converted VPD and Childsen to use the new prompt system.