Fix bug with Perilous Position (#532)
While implementing Bravado, I discovered there was a bug for Perilous
Position where if the unit was defeated by the upgrade, an exception
would be thrown when the upgrade tried to resolve the exhaust portion of
the ability. Since the unit was immediately defeated, when the exhaust
resolved the unit and upgrade would both already be in the discard. The
simplest solution was to remove the contract checks and let the ability
fizzle due to not having a target.

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>