Add CardDataGetter utility and consolidate testing + lobby flow to use it (#531)
Created a CardDataGetter base class that can be used for various
implementations of classes that retrieve our card data jsons for us.
This will make it much easier to make changes to how we get cards as
well as supporting alternate implementations - i.e., we currently have
one for downloading from AWS and one for pulling from local.

Consolidated the unit test harness and the GameServer + Lobby flows to
use this component. Did a lot of cleanup on the GameServer + Lobby to
improve the general flow and to use async correctly.