Refactor to use cardDataGetter and SWUDB decklist format (#599)
Previously, the FE would receive a SWUDB decklist and do the work of
pulling in all of the card data to create an internal decklist that had
all the data for each card.

To streamline things and make maintenance easier, we're migrating the
work of pulling card data to the BE. The FE will only need to pass in
the swudb decklist and the backend will consolidate it from there.

In addition, the card data is now being pulled in within the Game class
itself instead of being provided from outside. Unfortunately, this means
that our test harness now has to use an async setup method. Converted
all existing tests to use the new async approach.