Deck validation logic (#620)
Add in a basic deck validator for reporting issues such as bad deck
size, invalid deck format, illegal cards for format, or unimplemented
cards.

Right now this is kind of crude and definitely needs some refactoring to
support different formats better. Also needs a test suite. But dropping
it in now to unblock work on the FE side for integration.