Add Dooku leader and refactor exploit infrastructure (#655)
Added the Dooku leader. Refactored exploit to be a cost adjuster instead
of a subtype of the standard play cost so that it could be applied
dynamically to a play card action.

Did some general cleanup around exploit as well, including adding a more
explicit option to choose to play exploit and a cancel button (with some
TODOs for certain cases).


![image](https://github.com/user-attachments/assets/da31d6ca-6355-4fc5-939c-fe9afaccad2d)


![image](https://github.com/user-attachments/assets/785eb447-004a-4568-b5c0-f378c7d48771)