Relax MoveCard contract check (#828)
MoveCardSystem had a very strict contract check in canAffect making it
illegal to even ask if a card could be moved to an illegal zone. This
was triggering some hard-to-identify errors in the logs and is probably
too strong of a check anyway. Reduced it to just returning `false` in
that case.