Add Krennic unit and explicit registration for "when defeated" abilities (#893)
Added the Krennic unit. So support him and other cards that interact
with "when defeated" abilities, added a new type of ability registration
that allows explicitly registering common ability types like that one.
Also added a check in the card text itself to ensure that we always
register when defeated abilities correctly.


![image](https://github.com/user-attachments/assets/16e475d8-a889-4d80-9688-8fff6746e5cd)