Add cancel option for events (#966)
Made it possible to cancel out of the targeting stage for an event card.
Added a way for the the play event card step to do "very early"
targeting for the actual event ability and pass it forward. Also cleaned
up the "choose no target" button handling so that's more consistent.