Fix non-leader pilot bug and add better error handling for card condition cases (#974)
We had a bug where attached non-leader pilots were reporting true for
`isNonLeaderUnit()`, and with the recent fix their properties this meant
that things were trying to read their `damage` property and getting an
error.

With the recent change to error handling this causes a game state
failure. Added wrappers around the common pathways for checking whether
a card can be targeted by an effect so that if that check excepts
they'll just silently report an error and return "false" for that
target. In 99% of cases this should preserve game state correctly.