Card and Zone Initial State Changes  (#954)
#### Initial Card State Setup
I've setup basic state definitions for all of the card classes. A number
of the higher objects don't have state just overrides, so I was able to
skip those.

Mixins did make it harder, but if you pass in a state interface to the
initial class being mixed in it still can work.
```ts
const NonLeaderUnitCardParent = WithUnitProperties(WithStandardAbilitySetup(InPlayCard<INonLeaderUnitCardState>));
```

I kinda got the mixins themselves working to extend the state object
definition, but if you take a look at it, it's _ugly_. But it's the best
I could do without doing a deep dive into TS typings 😅.

I don't think I captured _all_ state for Cards, but this at least puts
the backbone in, so I don't have to worry about branch desyncs.

#### Game Object Manager
GameObjectManager's get method now throws an error if it is given a
valid obj/UUID but it can't find it. Because GameObjectRefs should only
ever be created from existing objects (ie they should only ever be
created from a obj.getRef() call), if the engine is given an UUID and
can't find it, something is fundamentally broken (something should be in
the state object but isn't) and we need to know ASAP.

Related, but the GameObjectRef system works flawlessly so far, so that's
good news.

#### Zone Changes
Zones are now derived from GameObjectBase. This means we can use
GameObjectRef<Zone> (see Card.ts for an example). Converted the two base
classes (SimpleZone and ConcreteArenaZone) that track cards to use the
state object.

ConcreteArenaZone had a noticeable rework, it now keeps all of the cards
in a list but now maintains the Map object as "cached" value. I added a
"onAfterStateSetup" method to GameObjectBase to call on objects after an
Undo is triggered, so that they can refresh any values that are
computations based on state (such as as the Map in this case). This does
mean ConcreteArenaZone is now maintaining two sets of objects, which is
not my favorite. I could probably add simplify this down further but
it's functional for now.

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>