Player and Game State changes and initial test undo setup (#999)
Got a round of state changes for Game and Player, as well as built a
initial, hacky way to test undo in test suites. Upon testing, I found
more state to hunt down, so I undid it and am putting this PR in next.

This primarily includes the state and functions to take and rollback
game objects to a snapshot in the (newly renamed) GameStateManager. I
started to work on moving initiative/active player from Game to the
manager but that started to get deep, so code is there in the manager
but not yet used.